chore: Remove typedoc dependency
