--- src/schema/index.ts
@@ -1,7 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-import * as code from './code';
-import * as types from './types';
-import * as utils from './utils';
-
-export default { code, types, utils };
