+++ test/test-utils/test-helpers.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { documentTestUtilsNew, TestUtilsVariantOptions } from '../../src/test-utils-new';
-import { TestUtilsDoc } from '../../src/test-utils-new/interfaces';
+import { documentTestUtilsNew, TestUtilsVariantOptions } from '../../src/test-utils';
+import { TestUtilsDoc } from '../../src/test-utils/interfaces';
 
 export function buildTestUtilsProject(
   name: string,
